Add customized TokenProvider for GitHub App
